+++ src/main/resources/db/scripts/008_ddl_create_car_table.sql
@@ -1,5 +1,5 @@
 CREATE TABLE car(
-    id serial primary key,
+    id bigserial primary key,
     name varchar not null,
-    engine_id int not null unique references engine(id)
+    engine_id bigint not null unique references engine(id)
 );
