+++ src/main/resources/db/scripts/009_ddl_create_history_owners_table.sql
@@ -1,7 +1,7 @@
 create table history_owners(
-    id       serial primary key,
-    owner_id int not null references owners (id),
-    car_id   int not null references car (id),
+    id       bigserial primary key,
+    owner_id bigint not null references owners (id),
+    car_id   bigint not null references car (id),
     startAt timestamp,
     endAt timestamp,
     UNIQUE (owner_id, car_id)
