+++ src/main/resources/db/scripts/010_ddl_alter_auto_post_table_add_car_id_col.sql
@@ -1,1 +1,1 @@
-ALTER TABLE auto_post ADD COLUMN car_id int REFERENCES car(id);
+ALTER TABLE auto_post ADD COLUMN car_id bigint REFERENCES car(id);
